% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_breach_depression_.R
\name{create_breach_depression}
\alias{create_breach_depression}
\title{Breach depressions}
\usage{
create_breach_depression(
  dem,
  output = tempfile(fileext = ".tif"),
  whitebox_wd = NULL,
  max_depth = Inf,
  max_length = Inf,
  flat_increment = NULL,
  fill_pits = FALSE,
  ...
)
}
\arguments{
\item{dem}{A character file path to a raster or a SpatRaster object.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}

\item{whitebox_wd}{valid working directory for whitebox to store temporary rasters. Defaults to NULL which stores in a temporary space that is deleted when the user session is over. You can use this to store rasters before they are processed into terra::rast objects. Otherwise it is suggested to leave this NULL and store the rasters using the output argument.}

\item{max_depth}{numeric, Optional maximum breach depth (default is \code{Inf}).}

\item{max_length}{numeric, Optional maximum breach channel length (in grid cells; default is \code{Inf}).}

\item{flat_increment}{numeric, Optional elevation increment applied to flat areas. Default is \code{NULL}.}

\item{fill_pits}{logical, Optional flag indicating whether to fill single-cell pits. Default is \code{FALSE}}

\item{...}{optional arguments passed to \code{whitebox::wbt()}}
}
\value{
A SpatRaster object.
}
\description{
Breaches all of the depressions in a DEM using Lindsay's (2016) algorithm. This should be preferred over depression filling in most cases.
}
\examples{
\dontshow{if (whitebox::check_whitebox_binary()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTR")
dem <- terra::rast(dem)
create_breach_depression(dem)
\dontshow{\}) # examplesIf}
}
