% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_breach_depression_lc.R
\name{create_breach_depression_lc}
\alias{create_breach_depression_lc}
\title{Use the least cost pathway method to breach depressions}
\usage{
create_breach_depression_lc(
  dem,
  output = tempfile(fileext = ".tif"),
  dist = 100,
  whitebox_wd = NULL,
  max_cost = NULL,
  min_dist = FALSE,
  flat_increment = NULL,
  fill = FALSE
)
}
\arguments{
\item{dem}{A character file path to a raster or a SpatRaster object.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}

\item{dist}{Maximum search distance for breach paths in cells. Defaults to 100.}

\item{whitebox_wd}{valid working directory for whitebox to store temporary rasters. Defaults to NULL which stores in a temporary space that is deleted when the user session is over. You can use this to store rasters before they are processed into terra::rast objects. Otherwise it is suggested to leave this NULL and store the rasters using the output argument.}

\item{max_cost}{numeric, Optional maximum breach cost (default is \code{Inf}).}

\item{min_dist}{logical, Optional flag indicating whether to minimize breach distances. Default is \code{FALSE}.}

\item{flat_increment}{Optional elevation increment applied to flat areas. Default is \code{NULL}.}

\item{fill}{logical, Optional flag indicating whether to fill any remaining unbreached depressions. Default is \code{FALSE}}
}
\value{
A SpatRaster object.
}
\description{
Creates a DEM with breached depressions using the least cost pathway. This is a
wrapper for \code{whitebox::wbt_breach_depressions_least_cost()} that accepts terra SpatRaster
objects or a file path to common raster files.
}
\examples{
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTR")
dem <- terra::rast(dem)
create_breach_depression_lc(dem, dist = 10)

}
