% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_D8_.R
\name{create_d8_fa}
\alias{create_d8_fa}
\title{Create a D8 flow accumulation raster}
\usage{
create_d8_fa(
  D8pointer,
  output = tempfile(fileext = ".tif"),
  whitebox_wd = NULL,
  out_type = "cells",
  log = FALSE,
  clip = FALSE,
  pntr = TRUE,
  esri_pntr = FALSE,
  ...
)
}
\arguments{
\item{D8pointer}{A character file path to a raster or a SpatRaster object.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}

\item{whitebox_wd}{valid working directory for whitebox to store temporary rasters. Defaults to NULL which stores in a temporary space that is deleted when the user session is over. You can use this to store rasters before they are processed into terra::rast objects. Otherwise it is suggested to leave this NULL and store the rasters using the output argument.}

\item{out_type}{A character value indicating the values calculated for each cell. One of \code{cells} (default), \verb{catchment area}, or \verb{specific contributing area}.}

\item{log}{Optional flag to request the output be log-transformed.}

\item{clip}{Optional flag to request clipping the display max by 1\%.}

\item{pntr}{Is the input raster a D8 flow pointer rather than a DEM?}

\item{esri_pntr}{Input D8 pointer uses the ESRI style scheme.}

\item{...}{optional arguments passed to \code{whitebox::wbt()}.}
}
\value{
A SpatRaster object.
}
\description{
Creates a flow accumulation grid from a D8 flow direction raster or imput DEM.
This is a wrapper for \code{whitebox::wbt_d8_flow_accumulation()} that accepts
terra SpatRaster objects or a file path to common raster files.
}
\examples{
\dontshow{if (whitebox::check_whitebox_binary()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTRdata")
dem <- terra::rast(dem)
D8pointer <- create_d8_pointer(dem)
create_d8_fa(D8pointer)
\dontshow{\}) # examplesIf}
}
