% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_D8.R
\name{create_d8_pointer}
\alias{create_d8_pointer}
\title{Use D8 algorithm to create flow direction grid.}
\usage{
create_d8_pointer(
  dem,
  output = tempfile(fileext = ".tif"),
  whitebox_wd = NULL,
  esri_pntr = FALSE,
  ...
)
}
\arguments{
\item{dem}{A character file path to a raster or a SpatRaster object.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}

\item{whitebox_wd}{valid working directory for whitebox to store temporary rasters. Defaults to NULL which stores in a temporary space that is deleted when the user session is over. You can use this to store rasters before they are processed into terra::rast objects. Otherwise it is suggested to leave this NULL and store the rasters using the output argument.}

\item{esri_pntr}{logical, D8 pointer uses the ESRI style scheme.}

\item{...}{optional arguments passed to \code{whitebox::wbt()}}
}
\value{
a SpatRaster object.
}
\description{
Creates a flow direction or pointer grid using the d8 algorithm. This is a
wrapper for the whitebox tools \code{d8_pointer} that accepts terra SpatRaster
objects or a file path to common raster files.
}
\examples{
\dontshow{if (whitebox::check_whitebox_binary()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTR")
dem <- terra::rast(dem)
create_d8_pointer(dem)
\dontshow{\}) # examplesIf}
}
