% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fill_depressions.R
\name{create_fill_depressions}
\alias{create_fill_depressions}
\title{Use Whitebox Tools fill depression algorithm.}
\usage{
create_fill_depressions(
  dem,
  output = tempfile(fileext = ".tif"),
  whitebox_wd = NULL,
  fix_flats = TRUE,
  flat_increment = NULL,
  max_depth = NULL,
  ...
)
}
\arguments{
\item{dem}{A character file path to a raster or a SpatRaster object.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}

\item{whitebox_wd}{valid working directory for whitebox to store temporary rasters. Defaults to NULL which stores in a temporary space that is deleted when the user session is over. You can use this to store rasters before they are processed into terra::rast objects. Otherwise it is suggested to leave this NULL and store the rasters using the output argument.}

\item{fix_flats}{logical, indicating whether flat areas should have a small gradient applied. Defaults to \code{TRUE}.}

\item{flat_increment}{Optional elevation increment applied to flat areas. Defaults to \code{NULL}.}

\item{max_depth}{Optional maximum depression depth to fill. Defaults to \code{NULL}.}

\item{...}{optional arguments passed to \code{whitebox::wbt()}}
}
\value{
A SpatRaster object.
}
\description{
Creates a DEM with all of the depressions filled and flat areas removed. This is a
wrapper for \code{whitebox::wbt_fill_depressions()} that accepts terra SpatRaster
objects or a file path to common raster files.
}
\examples{
\dontshow{if (whitebox::check_whitebox_binary()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTR")
dem <- terra::rast(dem)
create_fill_depressions(dem)
\dontshow{\}) # examplesIf}
}
