% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_single_pit_fill.R
\name{create_single_pit_fill}
\alias{create_single_pit_fill}
\title{Fill single cell pits}
\usage{
create_single_pit_fill(
  dem,
  output = tempfile(fileext = ".tif"),
  whitebox_wd = NULL,
  ...
)
}
\arguments{
\item{dem}{A character file path to a raster or a SpatRaster object.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}

\item{whitebox_wd}{A valid working directory for whitebox to store temporary rasters. Defaults to NULL which stores in a temporary space that is deleted when the user session is over. You can use this to store rasters before they are processed into terra::rast objects. Otherwise it is suggested to leave this NULL and store the rasters using the output argument.}

\item{...}{optional arguments passed to \code{whitebox::wbt()}}
}
\value{
A SpatRaster object
}
\description{
Wrapper for whitebox \code{FillSingleCellPits}. Raises pit cells to the elevation of their lowest neighbor.
}
\examples{
\dontshow{if (whitebox::check_whitebox_binary()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTRdata")
dem <- terra::rast(dem)
create_single_pit_fill(dem)
\dontshow{\}) # examplesIf}
}
