% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_streams.R
\name{create_streams}
\alias{create_streams}
\title{Extract stream network}
\usage{
create_streams(
  flow_accumulation,
  output = tempfile(fileext = ".tif"),
  whitebox_wd = NULL,
  threshold = 1000,
  zero_background = FALSE,
  ...
)
}
\arguments{
\item{flow_accumulation}{SpatRaster object, Input raster D8 flow accumulation file.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}

\item{whitebox_wd}{valid working directory for whitebox to store temporary rasters. Defaults to NULL which stores in a temporary space that is deleted when the user session is over. You can use this to store rasters before they are processed into terra::rast objects. Otherwise it is suggested to leave this NULL and store the rasters using the output argument.}

\item{threshold}{Numeric value indicating the threshold in flow accumulation values (number of cells) for channelization.}

\item{zero_background}{logical, Flag indicating whether a background value of zero should be used.}

\item{...}{optional arguments passed to \code{whitebox::wbt()}.}
}
\value{
A SpatRaster object.
}
\description{
Creates a raster of likely stream cells from a flow-accumulation raster using
the Whitebox Tools algorithm. This is a
wrapper for \code{whitebox::wbt_extract_streams()} that accepts terra SpatRaster
objects or a file path to common raster files.
}
\examples{
\dontshow{if (whitebox::check_whitebox_binary()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTRdata")
dem <- terra::rast(dem)
D8pointer <- create_d8_pointer(dem)
D8fa <- create_d8_fa(D8pointer)
create_streams(D8fa)
\dontshow{\}) # examplesIf}
}
