% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_streams_vector.R
\name{create_streams_vector}
\alias{create_streams_vector}
\title{Convert stream raster to vector}
\usage{
create_streams_vector(
  streams,
  d8_pointer,
  output = tempfile(fileext = ".shp"),
  whitebox_wd = NULL,
  esri_pntr = FALSE,
  all_vertices = FALSE
)
}
\arguments{
\item{streams}{SpatRaster object, input raster streams.}

\item{d8_pointer}{SpatRaster object, input raster D8 pointer.}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}

\item{whitebox_wd}{valid working directory for whitebox to store temporary rasters. Defaults to NULL which stores in a temporary space that is deleted when the user session is over. You can use this to store rasters before they are processed into terra::rast objects. Otherwise it is suggested to leave this NULL and store the rasters using the output argument.}

\item{esri_pntr}{logical, D8 pointer uses the ESRI style scheme.}

\item{all_vertices}{logical, Do you want to preserve all vertices in output (i.e. no straight-line generalization).}
}
\value{
A terra SpatVector object
}
\description{
Creates a stream vector (terra SpatVector object) from a D8 flow direction
raster and stream network raster.
This is a wrapper for \code{whitebox::wbt_raster_streams_to_vector()} that accepts
terra SpatRaster objects or a file path to common raster files.
}
\examples{
\dontshow{if (whitebox::check_whitebox_binary()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# example code
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTR")
dem <- terra::rast(dem)
D8pointer <- create_d8_pointer(dem)
D8fa <- create_d8_fa(D8pointer)
streams <- create_streams(D8fa)
create_streams_vector(streams, D8pointer)
\dontshow{\}) # examplesIf}
}
