% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_subbasins.R
\name{create_subbasins}
\alias{create_subbasins}
\title{Subbasins}
\usage{
create_subbasins(
  d8_pntr,
  streams,
  output = tempfile(fileext = ".tif"),
  esri_pntr = FALSE,
  whitebox_wd = NULL,
  type = "terra",
  ...
)
}
\arguments{
\item{d8_pntr}{Input D8 pointer raster file.}

\item{streams}{Input raster streams file.}

\item{output}{A character file path specifying where the shapefile file should be stored. Defaults to a temporary file.}

\item{esri_pntr}{D8 pointer uses the ESRI style scheme.}

\item{whitebox_wd}{valid working directory for whitebox to store temporary rasters. Defaults to NULL which stores in a temporary space that is deleted when the user session is over. You can use this to store rasters before they are processed into terra::rast objects. Otherwise it is suggested to leave this NULL and store the rasters using the output argument.}

\item{type}{character, one of \code{"terra"} or \code{"wbt"}. If \code{type = "terra"}, returns an object of SpatVector. If \code{type = "wbt"}, a wbt_result object is returned.}

\item{...}{optional arguments passed to \code{whitebox::wbt()}.}
}
\value{
SpatRaster
}
\description{
Identifies the catchments, or sub-basin, draining to each link in a stream network.
}
\examples{
\dontshow{if (whitebox::check_whitebox_binary()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTRdata")
dem <- terra::rast(dem)
D8pointer <- create_d8_pointer(dem)
D8fa <- create_d8_fa(D8pointer)
streams <- create_streams(D8fa)
create_subbasins(D8pointer, streams)
\dontshow{\}) # examplesIf}
}
