% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_streams.R
\name{extract_stream_outlets}
\alias{extract_stream_outlets}
\title{Extract last coordinate geom from stream vector}
\usage{
extract_stream_outlets(stream_vect, output = tempfile(fileext = ".shp"))
}
\arguments{
\item{stream_vect}{A SpatVector object of stream network}

\item{output}{A character file path specifying where the raster file should be stored. Defaults to a temporary file.}
}
\value{
A SpatVector points object for outlet stream points
}
\description{
Internal function to retrieve outlet coordinates for stream network vector.
}
\keyword{internal}
