% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SELECTRshed-package.R
\docType{package}
\name{SELECTRshed-package}
\alias{SELECTRshed}
\alias{SELECTRshed-package}
\title{SELECTRshed: Catchment and Subcatchment Delineation}
\description{
Use `terra` SpatRaster and SpatVector objects directly with WhiteBox Tools' stream and hydrology functions.
}
\author{
\strong{Maintainer}: Michael Schramm \email{mpschramm@gmail.com} (\href{https://orcid.org/0000-0003-1876-6592}{ORCID})

}
\keyword{internal}
