% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_pour_point.R
\name{snap_pour_point}
\alias{snap_pour_point}
\title{Jenson snap pour points}
\usage{
snap_pour_point(
  pour_pts,
  streams,
  output = tempfile(fileext = ".shp"),
  whitebox_wd = NULL,
  snap_dist = 90,
  type = "terra",
  ...
)
}
\arguments{
\item{pour_pts}{Input vector pour points (outlet) file.}

\item{streams}{Input raster streams file.}

\item{output}{A character file path specifying where the shapefile file should be stored. Defaults to a temporary file.}

\item{whitebox_wd}{valid working directory for whitebox to store temporary rasters. Defaults to NULL which stores in a temporary space that is deleted when the user session is over. You can use this to store rasters before they are processed into terra::rast objects. Otherwise it is suggested to leave this NULL and store the rasters using the output argument.}

\item{snap_dist}{Maximum snap distance in map units.}

\item{type}{character, one of \code{"terra"} or \code{"wbt"}. If \code{type = "terra"}, returns an object of SpatVector. If \code{type = "wbt"}, a wbt_result object is returned.}

\item{...}{optional arguments passed to \code{whitebox::wbt()}.}
}
\value{
terra SpatVector object.
}
\description{
Moves outlet points used to specify points of interest in a watershedding operation to the nearest stream cell.
}
